+++ src/ast/expr/mod.rs
@@ -1,9 +1,11 @@
 mod expr;
 mod literal;
 mod operator;
+mod syscall;
 mod unary;
 
 pub use expr::*;
 pub use literal::*;
 pub use operator::*;
+pub use syscall::*;
 pub use unary::*;
